Add custom entity mapper **DRY**
As we'll be adding more entities to our App, we use a custom
entity mapper to easily build entities from database records.
As per assignment definition and requirements, we can not use
an ORM; so, this helps us to reduce boilerplate code and makes
codebase easier to maintain.

Task: explore mapping alternatives for different App use-cases.